+++ apps/blog/src/app/not-found.tsx
@@ -1,10 +1,6 @@
 import { ArticleContainer } from '@components/article-container';
-import { findPosts } from '@libs/api/find-posts';
-import { findSeriesList } from '@libs/api/find-series';
-import { findTags } from '@libs/api/find-tags';
-import { Mapper } from '@libs/mapper';
+import { loadMainLayoutData } from '@libs/api/load-layout-data';
 import { PageLinkMap } from '@libs/page-link-map';
-import { PostModel } from '@libs/types/commons';
 import Link from 'next/link';
 import MainClientLayout from 'src/app/(main)/main-client-layout';
 
@@ -15,12 +11,7 @@ import MainClientLayout from 'src/app/(main)/main-client-layout';
  * MainClientLayout으로 감싼다. 막다른 길에서 홈·태그·최근 글로 복귀할 동선을 제공한다.
  */
 export default async function NotFound() {
-  const seriesModels = (await findSeriesList()).map(Mapper.toSeriesModel);
-  const tags = (await findTags()).map(Mapper.toTagModel);
-  const posts = (await findPosts())
-    .map((item) => Mapper.toPostModel({ item, seriesModels }))
-    // 시리즈 미존재로 스킵된 포스트(null)를 제거해 PostModel[]로 좁힌다.
-    .filter((post): post is PostModel => post !== null);
+  const { seriesModels, tags, posts } = await loadMainLayoutData();
 
   // 별도 인기도 데이터가 없어 최신 글 4개를 복구 동선으로 제시한다.
   const recentPosts = [...posts].sort((a, b) => b.date.localeCompare(a.date)).slice(0, 4);
